Remove macroquad dependency, replace it with SDL2

+++ src/display/consts.rs
@@ -1,6 +1,6 @@
-pub const FRAME_COLS: usize = 256;
-pub const FRAME_ROWS: usize = 240;
-pub const WINDOW_HEIGHT: usize = 480;
-pub const WINDOW_WIDTH: usize = 512;
-pub const PIXEL_WIDTH: f32 = WINDOW_WIDTH as f32 / FRAME_COLS as f32;
-pub const PIXEL_HEIGHT: f32 = WINDOW_HEIGHT as f32 / FRAME_ROWS as f32;
+pub const FRAME_COLS: u32 = 256;
+pub const FRAME_ROWS: u32 = 240;
+pub const WINDOW_HEIGHT: u32 = 480;
+pub const WINDOW_WIDTH: u32 = 512;
+pub const PIXEL_WIDTH: u32 = WINDOW_WIDTH / FRAME_COLS;
+pub const PIXEL_HEIGHT: u32 = WINDOW_HEIGHT / FRAME_ROWS;
